+++ Checkr/Repositories/Concrete/TagRepository.cs
@@ -9,41 +9,58 @@ namespace Checkr.Repositories.Concrete
     {
         private readonly ApplicationDbContext _context = context;
 
-        public List<Tag> GetAllTags()
+        public async Task<IEnumerable<Tag>> GetAllTagsForBoardAsync(int boardId)
         {
-            var tags = _context.Tags
+            return await _context.Tags
+                .Where(t => t.BoardId == boardId)
+                .Include(t => t.Board)
                 .Include(t => t.Boxes)
-                .ToList();
-            return tags;
+                .ToListAsync();
         }
 
-        public Tag GetTagById(int id)
+        public async Task<IEnumerable<Tag>> GetByIdsAsync(ICollection<int> ids)
         {
-            var tag = _context.Tags
+            return await _context.Tags
+                .Where(t => ids.Contains(t.Id))
+                .Include(t => t.Board)
                 .Include(t => t.Boxes)
-                .FirstOrDefault(l => l.Id == id);
+                .ToListAsync();
+        }
+
+        public async Task<Tag?> GetByIdAsync(int id)
+        {
+            return await _context.Tags
+                .Include(t => t.Board)
+                .Include(t => t.Boxes)
+                .FirstOrDefaultAsync(t => t.Id == id);
+        }
+
+        public async Task<Tag> CreateAsync(Tag tag)
+        {
+            await _context.Tags.AddAsync(tag);
+            await _context.SaveChangesAsync();
             return tag;
         }
 
-        public Tag AddTag(Tag tag)
-        {
-            _context.Tags.Add(tag);
-            _context.SaveChanges();
-            return tag;
-        }
-
-        public Tag UpdateTag(Tag tag)
+        public async Task<Tag> UpdateAsync(Tag tag)
         {
             _context.Tags.Update(tag);
-            _context.SaveChanges();
+            await _context.SaveChangesAsync();
             return tag;
         }
 
-        public Tag DeleteTag(int id)
+        public async Task<Tag?> DeleteAsync(int id)
         {
-            var tag = GetTagById(id);
-            _context.Tags.Remove(tag);
-            _context.SaveChanges();
+            var tag = await GetByIdAsync(id);
+            if (tag is not null)
+            {
+                foreach (var box in tag.Boxes)
+                {
+                    box.Tags.Remove(tag);
+                }
+                _context.Tags.Remove(tag);
+                await _context.SaveChangesAsync();
+            }
             return tag;
         }
     }
